+++ RELEASE_NOTES.md
@@ -5,7 +5,7 @@ This is the offical release of vpinfe v1.1.11.
 
 ### What's New
 - Revamp the main and collection menu cosmetically
-- Add table rating to mainmenu
+- Add table ratings to mainmenu / mangerui / Filtering 
 - Clean up theme.md of legacy
 
 ### Fixes